Fix build and test for older versions of node and bumped version.
The run sequence in gulp was different in older node versions.  This was actually a bug that did not show-up in later node versions.  Refactored to ensure things run in correct order.  Also, added extra code to skip optional modules that do not work in earlier nodes.  Print message to notify the user.